[frontend] fix crash after signup
- Remove proxy hack to signup and authenticate, this made express to lose the promise chain.
- use simple internal query.
- refactor internal query to return a simple Promise with data.
- Remove useless redirect on server side after signup.
- Change redirect on the client signup component.
- Refactor some js in ts.
