+++ src/frontend/web_application/package.json
@@ -125,6 +125,7 @@
     "@testing-library/user-event": "^13.1.9",
     "@types/classnames": "^2.3.1",
     "@types/express": "^4.17.13",
+    "@types/express-http-proxy": "^1.6.2",
     "@types/iron": "^5.0.2",
     "@types/jest": "^26.0.23",
     "@types/lingui__react": "^2.8.3",
