+++ src/frontend/web_application/server/error/consts.ts
@@ -1,0 +1,3 @@
+export class ServerError extends Error {
+  status = 500;
+}
